pyproject.toml
[project]
name = "app-open-source-recettes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
]
